+++ recipes/repository.rb
@@ -28,11 +28,16 @@ when 'debian'
     action :nothing
     subscribes :reload, 'package[lsb-release]', :immediately
   end
-
+  dist = node['lsb']['codename'] 
+  if platform?('ubuntu') && node['platform_version'].to_f >= 20.04
+        dist = node['lsb']['codename']  + '/' + node['packages']['apt']['arch'] + '/' 
+    elsif platform?('debian') && node['platform_version'].to_i >= 11
+        dist = node['lsb']['codename']  + '/' + node['packages']['apt']['arch'] + '/' 
+  end
   apt_repository 'ossec' do
-    uri 'http://updates.atomicorp.com/channels/atomic/' + node['platform']
-    key 'https://www.atomicorp.com/RPM-GPG-KEY.atomicorp.txt'
-    distribution lazy { node['lsb']['codename'] }
+    uri 'https://updates.atomicorp.com/channels/atomic/' + node['platform']
+    key 'https://updates.atomicorp.com/installers/RPM-GPG-KEY.atomicorp.txt'
+    distribution #{dist}
     components ['main']
   end
 end